move ckernel lrelu to relu, cut down templ params
