+++ tt_metal/hw/inc/api/compute/eltwise_unary/relu.h
@@ -7,9 +7,6 @@
 #include "api/compute/common_globals.h"
 #ifdef TRISC_MATH
 #include "ckernel_sfpu_relu.h"
-#ifdef ARCH_QUASAR
-#include "ckernel_sfpu_lrelu.h"
-#endif
 #include "llk_math_eltwise_unary_sfpu_macros.h"
 #endif
